Moved the Events to the parent model package as they are also relevant for all authentications
