+++ studymanager-auth-token/src/test/java/io/redlink/more/auth/token/service/LoginTokenServiceTest.java
@@ -9,8 +9,8 @@
 package io.redlink.more.auth.token.service;
 
 import io.redlink.more.auth.token.configuration.LoginTokenProperties;
-import io.redlink.more.auth.token.event.ParticipantUpdateAction;
-import io.redlink.more.auth.token.event.ParticipantUpdateEvent;
+import io.redlink.more.auth.event.ParticipantUpdateAction;
+import io.redlink.more.auth.event.ParticipantUpdateEvent;
 import io.redlink.more.auth.token.repository.LoginTokenRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
